Move event API to extension artifact
