+++ api/logs/src/main/java/io/opentelemetry/api/logs/LoggerProvider.java
@@ -5,7 +5,6 @@
 
 package io.opentelemetry.api.logs;
 
-import io.opentelemetry.api.events.EventLogger;
 import javax.annotation.concurrent.ThreadSafe;
 
 /**
@@ -17,7 +16,8 @@ import javax.annotation.concurrent.ThreadSafe;
  * <ol>
  *   <li>Enable emitting structured <a
  *       href="https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/logs/semantic_conventions/events.md">events</a>
- *       via {@link EventLogger#eventBuilder(String)}.
+ *       via {@code io.opentelemetry.api.events.EventLogger} in {@code
+ *       io.opentelemetry:opentelemetry-extension-events}.
  *   <li>Enable the creation of log appenders, which bridge logs from other log frameworks (e.g.
  *       SLF4J, Log4j, JUL, Logback, etc) into OpenTelemetry via {@link Logger#logRecordBuilder()}.
  *       It is <b>NOT</b> a replacement log framework.
