+++ extensions/events/README.md
@@ -1,0 +1,6 @@
+# OpenTelemetry Event API
+
+[![Javadocs][javadoc-image]][javadoc-url]
+
+[javadoc-image]: https://www.javadoc.io/badge/io.opentelemetry/opentelemetry-extension-events.svg
+[javadoc-url]: https://www.javadoc.io/doc/io.opentelemetry/opentelemetry-extension-events
