+++ extensions/events/src/main/java/io/opentelemetry/extension/event/EventBuilder.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.api.events;
+package io.opentelemetry.extension.event;
 
 import io.opentelemetry.api.logs.LogRecordBuilder;
