+++ extensions/events/src/test/java/io/opentelemetry/extension/event/EventLoggerTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.api.events;
+package io.opentelemetry.extension.event;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
