+++ settings.gradle.kts
@@ -31,6 +31,7 @@ include(":context")
 include(":dependencyManagement")
 include(":extensions:incubator")
 include(":extensions:aws")
+include(":extensions:events")
 include(":extensions:kotlin")
 include(":extensions:trace-propagators")
 include(":exporters:common")
